Rename components for clarity
